Aggiorna le funzioni di somma prefissata per supportare un numero maggiore di elementi e migliora la gestione degli errori nei shader

+++ include/prefixSum.hpp
@@ -19,14 +19,14 @@ void prefixSumMultiLevel1M(
 );
 
 void prefixSumMultiLevel1B(
-    GLuint countBuffer,          // input buffer with original counts
-    GLuint prefixSumBuffer,      // output prefix sum buffer (size = totalElements + 1)
-    GLuint blockSumsBuffer,      // temp: level 0 block sums
-    GLuint blockOffsetsBuffer,   // temp: level 0 block offsets
-    GLuint errorFlagBuffer,      // optional error flag buffer
-    Shader* prefixPass1,         // Level 0 prefix + block sums
-    Shader* prefixPass2,         // Higher level prefix (scanning block sums)
-    Shader* prefixPass3,         // Add block offsets to level 0 output
-    Shader* pass3BlockLevel,     // Add block offsets to block sums at levels > 1
-    int totalElements
-) ;
+    GLuint countBuffer,
+    GLuint prefixSumBuffer,
+    GLuint blockSumsBuffer,
+    GLuint blockOffsetsBuffer,
+    GLuint errorFlagBuffer,
+    Shader* prefixPass1,
+    Shader* prefixPass2,
+    Shader* prefixPass3,
+    Shader* pass3BlockLevel,
+    size_t totalElements
+);
